.cpp/.h turn to ./src
