--- announceManagement.h
@@ -1,10 +1,0 @@
-//
-// Created by ZGCC on 24-12-16.
-//
-
-#ifndef CPP_SE_ANNOUNCEMANAGEMENT_H
-#define CPP_SE_ANNOUNCEMANAGEMENT_H
-
-void AnnounceManagementGraph(pqxx::connection& conn);
-
-#endif //CPP_SE_ANNOUNCEMANAGEMENT_H
