--- cmake-build-debug/.cmake/api/v1/reply/target-CPP_SE-Debug-0dcaa51d8b90c2ed98a7.json
@@ -1,297 +1,0 @@
-{
-	"artifacts" : 
-	[
-		{
-			"path" : "CPP_SE.exe"
-		},
-		{
-			"path" : "CPP_SE.pdb"
-		}
-	],
-	"backtrace" : 1,
-	"backtraceGraph" : 
-	{
-		"commands" : 
-		[
-			"add_executable",
-			"link_directories",
-			"target_link_libraries",
-			"library_target_setup",
-			"include_directories"
-		],
-		"files" : 
-		[
-			"CMakeLists.txt",
-			"libpqxx/src/CMakeLists.txt"
-		],
-		"nodes" : 
-		[
-			{
-				"file" : 0
-			},
-			{
-				"command" : 0,
-				"file" : 0,
-				"line" : 15,
-				"parent" : 0
-			},
-			{
-				"command" : 1,
-				"file" : 0,
-				"line" : 7,
-				"parent" : 0
-			},
-			{
-				"command" : 2,
-				"file" : 0,
-				"line" : 27,
-				"parent" : 0
-			},
-			{
-				"file" : 1
-			},
-			{
-				"command" : 3,
-				"file" : 1,
-				"line" : 81,
-				"parent" : 4
-			},
-			{
-				"command" : 2,
-				"file" : 1,
-				"line" : 29,
-				"parent" : 5
-			},
-			{
-				"command" : 2,
-				"file" : 1,
-				"line" : 31,
-				"parent" : 5
-			},
-			{
-				"command" : 4,
-				"file" : 0,
-				"line" : 6,
-				"parent" : 0
-			}
-		]
-	},
-	"compileGroups" : 
-	[
-		{
-			"compileCommandFragments" : 
-			[
-				{
-					"fragment" : "-g -std=gnu++20 -fdiagnostics-color=always"
-				}
-			],
-			"includes" : 
-			[
-				{
-					"backtrace" : 8,
-					"path" : "F:/JetBrains/CPP_SE/./easyx/include"
-				},
-				{
-					"backtrace" : 3,
-					"path" : "F:/JetBrains/CPP_SE/cmake-build-debug/build-libpqxx/include"
-				},
-				{
-					"backtrace" : 3,
-					"path" : "F:/JetBrains/CPP_SE/libpqxx/include"
-				}
-			],
-			"language" : "CXX",
-			"languageStandard" : 
-			{
-				"backtraces" : 
-				[
-					1
-				],
-				"standard" : "20"
-			},
-			"sourceIndexes" : 
-			[
-				0,
-				2,
-				4,
-				6,
-				8,
-				10
-			]
-		}
-	],
-	"dependencies" : 
-	[
-		{
-			"backtrace" : 3,
-			"id" : "pqxx::@c43edb16e53f8d75f02f"
-		}
-	],
-	"id" : "CPP_SE::@6890427a1f51a3e7e1df",
-	"link" : 
-	{
-		"commandFragments" : 
-		[
-			{
-				"fragment" : "-g",
-				"role" : "flags"
-			},
-			{
-				"fragment" : "",
-				"role" : "flags"
-			},
-			{
-				"backtrace" : 2,
-				"fragment" : "-LF:\\JetBrains\\CPP_SE\\.\\easyx\\lib64",
-				"role" : "libraryPath"
-			},
-			{
-				"fragment" : "-Wl,-Bstatic",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 1,
-				"fragment" : "-leasyx",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 3,
-				"fragment" : "build-libpqxx\\src\\libpqxx.a",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 1,
-				"fragment" : "-leasyx",
-				"role" : "libraries"
-			},
-			{
-				"fragment" : "-Wl,-Bdynamic",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 6,
-				"fragment" : "\"C:\\Program Files\\PostgreSQL\\17\\lib\\libpq.lib\"",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 7,
-				"fragment" : "-lwsock32",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 7,
-				"fragment" : "-lws2_32",
-				"role" : "libraries"
-			},
-			{
-				"fragment" : "-lkernel32 -luser32 -lgdi32 -lwinspool -lshell32 -lole32 -loleaut32 -luuid -lcomdlg32 -ladvapi32",
-				"role" : "libraries"
-			}
-		],
-		"language" : "CXX"
-	},
-	"name" : "CPP_SE",
-	"nameOnDisk" : "CPP_SE.exe",
-	"paths" : 
-	{
-		"build" : ".",
-		"source" : "."
-	},
-	"sourceGroups" : 
-	[
-		{
-			"name" : "Source Files",
-			"sourceIndexes" : 
-			[
-				0,
-				2,
-				4,
-				6,
-				8,
-				10
-			]
-		},
-		{
-			"name" : "Header Files",
-			"sourceIndexes" : 
-			[
-				1,
-				3,
-				5,
-				7,
-				9,
-				11
-			]
-		}
-	],
-	"sources" : 
-	[
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "main.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "ustil.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "graphFunctions.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "graphFunctions.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "studentManagement.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "studentManagement.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "teacherManagement.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "teacherManagement.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "userManagement.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "userManagement.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "announceManagement.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "announceManagement.h",
-			"sourceGroupIndex" : 1
-		}
-	],
-	"type" : "EXECUTABLE"
-}
