+++ cmake-build-debug/.ninja_log
@@ -1,12 +1,13 @@
 # ninja log v5
-329	4856	7560676065103007	CMakeFiles/CPP_SE.dir/teacherManagement.cpp.obj	a9ed2cccfe904978
+10	4072	7560691255196861	CMakeFiles/CPP_SE.dir/teacherManagement.cpp.obj	a9ed2cccfe904978
 330	4756	7560422962343605	build-libpqxx/src/CMakeFiles/pqxx.dir/blob.cxx.obj	f66e67b0b419decf
-8	3529	7560678867682373	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
+13	5070	7560695461495176	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
 1427	5534	7560422974424263	build-libpqxx/src/CMakeFiles/pqxx.dir/largeobject.cxx.obj	f97a8e5e8a1bfa94
-4	32808	7560547867366809	build.ninja	a5ecc825e26b2dc8
+4	32808	7560724370869904	build.ninja	a5ecc825e26b2dc8
 650	4953	7560422963526368	build-libpqxx/src/CMakeFiles/pqxx.dir/cursor.cxx.obj	94c372c0a87beeef
 202	3741	7560422956865695	build-libpqxx/src/CMakeFiles/pqxx.dir/binarystring.cxx.obj	c40db399577ee77d
 1135	4588	7560422960787999	build-libpqxx/src/CMakeFiles/pqxx.dir/except.cxx.obj	a8efa62216c5d3f
+516	7260	7560722678852304	CMakeFiles/CPP_SE.dir/JetBrains/CPP/GoodsQuery.cpp.obj	2c70eebc5ecaecb8
 88	3948	7560422958443405	build-libpqxx/src/CMakeFiles/pqxx.dir/array.cxx.obj	5b480c8c20b885de
 811	4219	7560422958800094	build-libpqxx/src/CMakeFiles/pqxx.dir/encodings.cxx.obj	551ff1e9dd9bf49
 994	4406	7560422961399063	build-libpqxx/src/CMakeFiles/pqxx.dir/errorhandler.cxx.obj	caa7a92f750a2cd0
@@ -16,6 +17,7 @@
 8308	11406	7560423034332965	build-libpqxx/src/CMakeFiles/pqxx.dir/wait.cxx.obj	834d5f584af91b0a
 1285	5150	7560422965881332	build-libpqxx/src/CMakeFiles/pqxx.dir/field.cxx.obj	78161a442248e581
 1593	5336	7560422971479380	build-libpqxx/src/CMakeFiles/pqxx.dir/notification.cxx.obj	c7646f49e5377074
+1437	7109	7560722677339785	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/announceManagement.cpp.obj	f616f4f77a741047
 1994	6745	7560422986401954	build-libpqxx/src/CMakeFiles/pqxx.dir/pipeline.cxx.obj	77be19c99f91a33f
 1786	7017	7560422989252905	build-libpqxx/src/CMakeFiles/pqxx.dir/params.cxx.obj	9d7a975c83cfadca
 3742	8160	7560423000643502	build-libpqxx/src/CMakeFiles/pqxx.dir/result.cxx.obj	d7e7867e77d77519
@@ -23,39 +25,34 @@
 4588	8797	7560423007725057	build-libpqxx/src/CMakeFiles/pqxx.dir/strconv.cxx.obj	6ea4f36ff0c7d3be
 4406	8503	7560423003691182	build-libpqxx/src/CMakeFiles/pqxx.dir/sql_cursor.cxx.obj	df3cb9beed50b06c
 8160	8873	7560423008900513	build-libpqxx/src/CMakeFiles/pqxx.dir/version.cxx.obj	be43231d13861995
-452	4908	7560676065611274	CMakeFiles/CPP_SE.dir/userManagement.cpp.obj	f93d4c4ddfca2bca
+1873	7602	7560722682294706	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/graphFunctions.cpp.obj	997d369474afb28f
+138	4137	7560691255833552	CMakeFiles/CPP_SE.dir/userManagement.cpp.obj	f93d4c4ddfca2bca
 489	8825	7560423007800245	build-libpqxx/src/CMakeFiles/pqxx.dir/connection.cxx.obj	eeece00790f6b401
-579	4975	7560676066300367	CMakeFiles/CPP_SE.dir/announceManagement.cpp.obj	2ea42f636529fa58
+283	4215	7560691256622135	CMakeFiles/CPP_SE.dir/announceManagement.cpp.obj	2ea42f636529fa58
 4756	8849	7560423008289906	build-libpqxx/src/CMakeFiles/pqxx.dir/stream_from.cxx.obj	8dda43c510602ab4
 5534	9146	7560423011425430	build-libpqxx/src/CMakeFiles/pqxx.dir/transaction.cxx.obj	dfacfc6b484a41d8
 5336	9255	7560423012456306	build-libpqxx/src/CMakeFiles/pqxx.dir/time.cxx.obj	1f7267b5a5c2b1e4
 7017	9718	7560423017313152	build-libpqxx/src/CMakeFiles/pqxx.dir/util.cxx.obj	517900012cd1ed9c
 6745	9857	7560423018790396	build-libpqxx/src/CMakeFiles/pqxx.dir/transaction_base.cxx.obj	3e1395746f0b7c2c
+2163	7833	7560722684603095	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/studentManagement.cpp.obj	ecc4688f47bf0e3
 141	1576	7560425088835370	build-libpqxx/src/libpqxx.a	c8ad7a32652798f2
-3530	4354	7560678875843818	CPP_SE.exe	61ee15e280e9c50a
-97	4542	7560676061948896	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
-215	4677	7560676063310655	CMakeFiles/CPP_SE.dir/studentManagement.cpp.obj	47696220330eb3aa
-9	3548	7560679006485786	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
-3549	4363	7560679014564371	CPP_SE.exe	61ee15e280e9c50a
-108	4715	7560690927793696	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
-234	4893	7560690929566777	CMakeFiles/CPP_SE.dir/studentManagement.cpp.obj	47696220330eb3aa
-400	4990	7560690930500260	CMakeFiles/CPP_SE.dir/teacherManagement.cpp.obj	a9ed2cccfe904978
-9	5051	7560690931119366	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
-517	5177	7560690932387366	CMakeFiles/CPP_SE.dir/userManagement.cpp.obj	f93d4c4ddfca2bca
-655	5363	7560690934253870	CMakeFiles/CPP_SE.dir/announceManagement.cpp.obj	2ea42f636529fa58
-10	4072	7560691255196861	CMakeFiles/CPP_SE.dir/teacherManagement.cpp.obj	a9ed2cccfe904978
-138	4137	7560691255833552	CMakeFiles/CPP_SE.dir/userManagement.cpp.obj	f93d4c4ddfca2bca
-283	4215	7560691256622135	CMakeFiles/CPP_SE.dir/announceManagement.cpp.obj	2ea42f636529fa58
-4215	5404	7560691268408104	CPP_SE.exe	61ee15e280e9c50a
-117	3929	7560691786462114	CMakeFiles/CPP_SE.dir/studentManagement.cpp.obj	47696220330eb3aa
-9	4272	7560691789729098	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
-4273	5418	7560691801313215	CPP_SE.exe	61ee15e280e9c50a
-8	3538	7560692193562848	CMakeFiles/CPP_SE.dir/studentManagement.cpp.obj	47696220330eb3aa
-3538	4688	7560692204989634	CPP_SE.exe	61ee15e280e9c50a
-10	4185	7560692593130685	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
-4185	5409	7560692605286349	CPP_SE.exe	61ee15e280e9c50a
-8	4246	7560693541911712	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
-4247	5739	7560693553975228	CPP_SE.exe	61ee15e280e9c50a
-109	4717	7560695457268878	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
-13	5070	7560695461495176	CMakeFiles/CPP_SE.dir/main.cpp.obj	b80ff7edd6443bc2
+552	5095	7560723738495508	CMakeFiles/CPP_SE.dir/src/teacherManagement.cpp.obj	422175defabfbeed
 5070	6394	7560695474637208	CPP_SE.exe	61ee15e280e9c50a
+109	4717	7560695457268878	CMakeFiles/CPP_SE.dir/graphFunctions.cpp.obj	3217253242501fab
+1732	8839	7560722694655209	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/example.cpp.obj	9266d93172893ac9
+9	4510	7560723732630150	CMakeFiles/CPP_SE.dir/src/announceManagement.cpp.obj	b9e3ed3552925364
+435	5009	7560723737629092	CMakeFiles/CPP_SE.dir/src/studentManagement.cpp.obj	631123d094765d62
+8	3538	7560692193562848	CMakeFiles/CPP_SE.dir/studentManagement.cpp.obj	47696220330eb3aa
+915	8209	7560722688372347	CMakeFiles/CPP_SE.dir/JetBrains/CPP/example.cpp.obj	a1cf220e4f2d4187
+781	1872	7560722618904778	CMakeFiles/CPP_SE.dir/JetBrains/CPP/cmake-build-debug/CMakeFiles/3.28.1/CompilerIdCXX/CMakeCXXCompilerId.cpp.obj	17d02d9cf569278a
+1168	2014	7560722623295974	CMakeFiles/CPP_SE.dir/JetBrains/CPP20/cmake-build-debug/CMakeFiles/3.28.1/CompilerIdCXX/CMakeCXXCompilerId.cpp.obj	9c1aa712e0d0e50d
+1567	2163	7560722627455646	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/cmake-build-debug/CMakeFiles/3.28.1/CompilerIdCXX/CMakeCXXCompilerId.cpp.obj	771acbb26c65571c
+1305	5319	7560722659467104	CMakeFiles/CPP_SE.dir/JetBrains/CPP20/main.cpp.obj	ec5ae6c507c45ec4
+651	7653	7560722682783287	CMakeFiles/CPP_SE.dir/JetBrains/CPP/UserQuery.cpp.obj	b12e921ca4e2d2c2
+1038	7940	7560722685663043	CMakeFiles/CPP_SE.dir/JetBrains/CPP/main.cpp.obj	2d3bec60cc23f020
+2014	8248	7560722688738568	CMakeFiles/CPP_SE.dir/JetBrains/CPP_SE/main.cpp.obj	d80adbb85c298bf3
+206	4708	7560723734621050	CMakeFiles/CPP_SE.dir/src/graphFunctions.cpp.obj	103049ebbb3423c0
+675	5223	7560723739765841	CMakeFiles/CPP_SE.dir/src/userManagement.cpp.obj	7140d08237843fe3
+317	5360	7560723741094692	CMakeFiles/CPP_SE.dir/src/main.cpp.obj	ceac5b88e4576b9b
+98	5886	7560723746394952	CMakeFiles/CPP_SE.dir/src/example.cpp.obj	4dc061c994ff9973
+8	1143	7560724937219194	CPP_SE.exe	9172dd0af4185839
