+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,6 +1,6 @@
 "D:\Jetbrain\CLion 2024.1.1\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug "-DCMAKE_MAKE_PROGRAM=D:/Jetbrain/CLion 2024.1.1/bin/ninja/win/x64/ninja.exe" -G Ninja -S F:\JetBrains\CPP_SE -B F:\JetBrains\CPP_SE\cmake-build-debug
 -- Generating config.h
 -- Generating config.h - done
--- Configuring done (36.1s)
+-- Configuring done (33.4s)
 -- Generating done (0.0s)
 -- Build files have been written to: F:/JetBrains/CPP_SE/cmake-build-debug
