+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Dec 17 00:50 中国标准时间
+Start testing: Dec 17 01:14 中国标准时间
 ----------------------------------------------------------
-End testing: Dec 17 00:50 中国标准时间
+End testing: Dec 17 01:14 中国标准时间
