+++ src/teacherManagement.h
@@ -1,0 +1,10 @@
+//
+// Created by ZGCC on 24-12-16.
+//
+
+#ifndef CPP_SE_TEACHERMANAGEMENT_H
+#define CPP_SE_TEACHERMANAGEMENT_H
+
+void TeacherManagementGraph(pqxx::connection& conn);
+
+#endif //CPP_SE_TEACHERMANAGEMENT_H
