+++ src/userManagement.h
@@ -1,0 +1,12 @@
+//
+// Created by ZGCC on 24-12-16.
+//
+
+#ifndef CPP_SE_USERMANAGEMENT_H
+#define CPP_SE_USERMANAGEMENT_H
+
+
+
+void UserManagementGraph(pqxx::connection& conn);
+
+#endif //CPP_SE_USERMANAGEMENT_H
