--- userManagement.h
@@ -1,12 +1,0 @@
-//
-// Created by ZGCC on 24-12-16.
-//
-
-#ifndef CPP_SE_USERMANAGEMENT_H
-#define CPP_SE_USERMANAGEMENT_H
-
-
-
-void UserManagementGraph(pqxx::connection& conn);
-
-#endif //CPP_SE_USERMANAGEMENT_H
